refactor(protocol): extract cryptographic functions and signatures into dedicated modules; add a seperate model/tokenizer loader!
